requirements.txt
torch>=1.10.0
torchvision>=0.11.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.4.0
streamlit>=1.0.0
tensorboard>=2.7.0
